% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kll.R
\docType{methods}
\name{approx_cdf}
\alias{approx_cdf}
\alias{approx_cdf,numeric,integer-method}
\alias{approx_cdf,DelayedArray,integer-method}
\title{Compute the approximate CDF of some numeric values.}
\usage{
approx_cdf(object, k, ...)

\S4method{approx_cdf}{numeric,integer}(object, k)

\S4method{approx_cdf}{DelayedArray,integer}(object, k)
}
\arguments{
\item{object}{A collection of numeric values.}

\item{k}{The K parameter for creating a KLL sketch.}

\item{...}{Extra named arguments.}
}
\description{
Compute the approximate CDF of some numeric values.
}
