% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\docType{methods}
\name{approx_col_cdf}
\alias{approx_col_cdf}
\alias{approx_col_cdf,matrix,integer-method}
\alias{approx_col_cdf,DelayedMatrix,integer-method}
\title{Compute the approximate CDF of the columns in a numeric array.}
\usage{
approx_col_cdf(matrix, k, ...)

\S4method{approx_col_cdf}{matrix,integer}(matrix, k)

\S4method{approx_col_cdf}{DelayedMatrix,integer}(matrix, k)
}
\arguments{
\item{matrix}{A numerical matrix.}

\item{k}{The K parameter for creating a KLL sketch.}

\item{...}{Extra named arguments.}
}
\description{
Compute the approximate CDF of the columns in a numeric array.
}
