% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{kll}
\alias{kll}
\alias{kll-package}
\title{kll}
\description{
An implementation of first algorithm from Karnin, Lang and Liberty
for computing approximate quantiles in the streaming computational
model.
}
\author{
Gabriele Sales
}
