% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kll.R
\docType{methods}
\name{kll_cdf}
\alias{kll_cdf}
\alias{kll_cdf,numeric,integer-method}
\alias{kll_cdf,DelayedArray,integer-method}
\title{Compute the CDF of some numeric values.}
\usage{
kll_cdf(object, k, ...)

\S4method{kll_cdf}{numeric,integer}(object, k)

\S4method{kll_cdf}{DelayedArray,integer}(object, k)
}
\arguments{
\item{object}{A collection of numeric values.}

\item{k}{The K parameter for creating a KLL sketch.}

\item{...}{Extra named arguments.}
}
\description{
Compute the CDF of some numeric values.
}
